A more advanced mechanism to keep track of the state of open sockets.
The purpose is to fix a kind of situations where a socket would be left open
with data still coming in from the Lisp side while the handler context would
be already discarded (presumably, due to a combination of keep-alive and
interrupted loading).  This would cause lighttpd to either crash or hang. 

Under the new mechanism, there is a pool from which socket slots are
allocated to handler instances.  A slot contains the socket fd together
with some metadata.  Instead of a single UnsafeLispSocket flag it contains
a state indicator/counter which is incremented for any event which could
not be handled by handle_fdevent.  Slots with sufficiently large unhandled
event counts are returned into the pool (as are slots with closed sockets).

The fixed size of the pool may be also used to limit the number of
simultaneous connections to the Lisp side.